sparse batched matrix fwt2d passes test.
