+++ src/ptwt/matmul_transform_2d.py
@@ -3,12 +3,7 @@
 import torch
 import numpy as np
 
-
-from .sparse_math import (
-    sparse_kron,
-    sparse_diag,
-    _dense_kron
-)
+from .sparse_math import construct_strided_conv2d_matrix
 from .conv_transform import (
     flatten_2d_coeff_lst,
     construct_2d_filt,
@@ -18,184 +13,6 @@ from .matmul_transform import (
     orthogonalize,
     cat_sparse_identity_matrix
 )
-
-
-def construct_conv_matrix(filter: torch.tensor,
-                          input_columns: int,
-                          mode: str = 'valid') -> torch.Tensor:
-    """Constructs a convolution matrix,
-       full and valid padding are supported.
-
-    Args:
-        filter (torch.tensor): The 1d-filter to convolve with.
-        input_columns (int): The number of columns in the input.
-        mode (str): String indetifier for the desired padding.
-            Defaults to valid.
-
-    Returns:
-        torch.Tensor: The sparse convolution tensor.
-
-    For reference see:
-    https://github.com/RoyiAvital/StackExchangeCodes/blob/\
-        master/StackOverflow/Q2080835/CreateConvMtxSparse.m
-    """
-    filter_length = len(filter)
-
-    if mode == 'full':
-        start_row = 0
-        stop_row = input_columns + filter_length - 1
-    elif mode == 'same' or mode == 'sameshift':
-        filter_offset = filter_length % 2
-        # signal_offset = input_columns % 2
-        start_row = filter_length // 2 - 1 + filter_offset
-        stop_row = start_row + input_columns - 1
-    elif mode == 'valid':
-        start_row = filter_length - 1
-        stop_row = input_columns - 1
-    else:
-        raise ValueError('unkown padding type.')
-
-    row_indices = []
-    column_indices = []
-    values = []
-    for column in range(0, input_columns):
-        for row in range(0, filter_length):
-            check_row = column + row
-            if (check_row >= start_row) and (check_row <= stop_row):
-                row_indices.append(row + column - start_row)
-                column_indices.append(column)
-                values.append(filter[row])
-    indices = np.stack([row_indices, column_indices])
-    values = torch.stack(values)
-    return torch.sparse_coo_tensor(indices, values, dtype=filter.dtype)
-
-
-def construct_conv2d_matrix(filter: torch.tensor,
-                            input_rows: int,
-                            input_columns: int,
-                            mode: str = 'valid',
-                            fully_sparse: bool = True) -> torch.Tensor:
-    """ Create a two dimensional sparse convolution matrix.
-        Convolving with this matrix should be equivalent to
-        a call to scipy.signal.convolve2d and a reshape.
-
-    Args:
-        filter (torch.tensor): A filter of shape [height, width]
-            to convolve with.
-        input_rows (int): The number of rows in the input matrix.
-        input_columns (int): The number of columns in the input matrix.
-        mode: (str) = The desired padding method. Options are
-            full, same and valid. Defaults to 'valid' or no padding.
-        fully_sparse (bool): Use a sparse implementation of the Kronecker
-            to save memory. Defaults to True.
-    Returns:
-        [torch.sparse.FloatTensor]: A sparse convolution matrix.
-    """
-    if fully_sparse:
-        kron = sparse_kron
-    else:
-        kron = _dense_kron
-
-    kernel_column_number = filter.shape[-1]
-    matrix_block_number = kernel_column_number
-
-    block_matrix_list = []
-    for i in range(matrix_block_number):
-        block_matrix_list.append(construct_conv_matrix(
-            filter[:, i], input_rows, mode))
-
-    if mode == 'full':
-        diag_index = 0
-        kronecker_rows = input_columns + kernel_column_number - 1
-    elif mode == 'same' or mode == 'sameshift':
-        filter_offset = kernel_column_number % 2
-        diag_index = kernel_column_number // 2 - 1 + filter_offset
-        kronecker_rows = input_columns
-    elif mode == 'valid':
-        diag_index = kernel_column_number - 1
-        kronecker_rows = input_columns - kernel_column_number + 1
-    else:
-        raise ValueError('unknown conv type.')
-
-    diag_values = torch.ones([int(np.min([kronecker_rows, input_columns]))],
-                             dtype=filter.dtype, device=filter.device)
-    diag = sparse_diag(diag_values, diag_index, kronecker_rows, input_columns)
-    sparse_conv_matrix = kron(diag, block_matrix_list[0])
-
-    for block_matrix in block_matrix_list[1:]:
-        diag_index -= 1
-        diag = sparse_diag(diag_values, diag_index,
-                           kronecker_rows, input_columns)
-        sparse_conv_matrix += kron(diag, block_matrix)
-
-    return sparse_conv_matrix
-
-
-def construct_strided_conv2d_matrix(
-        filter: torch.Tensor,
-        input_rows: int,
-        input_columns: int,
-        stride: int = 2,
-        mode='full') -> torch.Tensor:
-    """ Create a strided sparse two dimensional convolution
-       matrix.
-
-    Args:
-        filter (torch.tensor): The two dimensional convolution filter.
-        input_rows (int): The number of rows in the 2d-input matrix.
-        input_columns (int): The number of columns in the 2d- input matrix.
-        stride (int, optional): The stride between the filter positions.
-            Defaults to 2.
-        mode (str, optional): The convolution type.
-            Options are 'full', 'valid', 'same' and 'sameshift'.
-            Defaults to 'full'.
-
-    Raises:
-        ValueError: Raised if an unknown convolution string is
-            provided.
-
-    Returns:
-        [torch.Tensor]: The sparse convolution tensor.
-    """
-    filter_shape = filter.shape
-    convolution_matrix = construct_conv2d_matrix(
-        filter,
-        input_rows, input_columns, mode=mode)
-
-    if mode == 'full':
-        output_rows = filter_shape[0] + input_rows - 1
-        output_columns = filter_shape[1] + input_columns - 1
-    elif mode == 'valid':
-        output_rows = input_rows - filter_shape[0] + 1
-        output_columns = input_columns - filter_shape[1] + 1
-    elif mode == 'same' or mode == 'sameshift':
-        output_rows = input_rows
-        output_columns = input_columns
-    else:
-        raise ValueError("Padding mode not accepted.")
-
-    output_elements = output_rows * output_columns
-    element_numbers = torch.arange(output_elements,
-                                   device=filter.device).reshape(
-        output_columns, output_rows)
-
-    start = 0
-    if mode == 'sameshift':
-        start += 1
-
-    strided_rows = element_numbers[start::stride, start::stride]
-    strided_rows = strided_rows.flatten()
-    selection_eye = torch.sparse_coo_tensor(
-        torch.stack([torch.arange(len(strided_rows),
-                                  device=convolution_matrix.device),
-                     strided_rows],
-                    0),
-        torch.ones(len(strided_rows)),
-        dtype=convolution_matrix.dtype,
-        device=convolution_matrix.device,
-        size=[len(strided_rows), convolution_matrix.shape[0]])
-    # return convolution_matrix.index_select(0, strided_rows) 
-    return torch.sparse.mm(selection_eye, convolution_matrix)
 
 
 def construct_a_2d(wavelet, height: int, width: int,
@@ -396,9 +213,9 @@ class MatrixWavedec2d(object):
             split_size = int(np.prod(size))
             four_split = torch.split(next_to_split, split_size)
             next_to_split = four_split[0]
-            reshaped = tuple(torch.reshape(el.T, (batch_size, size[0], size[1])) for el in four_split[1:])
+            reshaped = tuple((el.T.reshape(batch_size, size[0], size[1])) for el in four_split[1:])
             split_list.append(reshaped)
-        split_list.append(torch.reshape(next_to_split, (batch_size, size[0], size[1])))
+        split_list.append(next_to_split.T.reshape(batch_size, size[0], size[1]))
 
         return split_list[::-1]
 
@@ -468,7 +285,7 @@ class MatrixWaverec2d(object):
         reconstruction = torch.sparse.mm(
             self.ifwt_matrix, coefficient_vectors.T)
 
-        return reconstruction.reshape((batch_size, height, width))
+        return reconstruction.T.reshape((batch_size, height, width))
 
 
 if __name__ == '__main__':
